Fix very slow shutdown which at times causes shutdown to be forced
replicationStream.read() is the cause of the problem due to it being
a blocking operation. Switch to replicationStream.readPending() which
is non-blocking. The polling thread will become very busy after making
this change, since it will be looping continuously - add a polling
interval delay after all data has been read to counter this.

Unrelated to the above, add a configuration option for being able to
set the `statusInterval` which determines when the LSN updates are
sent back to the server. This will hopefully allow the tests to be
less flaky on assertions based on LSN behaviour.